separate api-classes to diferent files with module
